store in its own folder and seperated out
