+++ src/components/MyStats.js
@@ -2,6 +2,9 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import GameCard from './GameCard';
 
+//so at this point we should have a user ID from being logged in
+// with the user id we can querry requests for all games played 
+// that will give us a list of games played that needs to be joined with 
 export class MyStats extends Component{
 
   render(){
@@ -39,6 +42,10 @@ const mapState = ({users, games, requests, match}) => {
   //need to setup url path for a specific user id
   const user = users.find( user => user.id === 4 ) || {};
   const request = requests.find( request => request.id === 24 ) || {};
+  
+  const request2 = requests.find( request => request.userId === user.id && request.gameId !== null ) || {};
+  
+  console.log(request2);
   const game = games.find( game => game.id * 1 === 1 ) || {};
 
   return {
