+++ src/store/users.js
@@ -5,16 +5,6 @@ const CREATE_USER = 'CREATE_USER';
 const DESTROY_USER = 'DESTROY_USER';
 const UPDATE_USER = 'UPDATE_USER';
 
-const LOAD_REQUESTS = 'LOAD_REQUESTS';
-const CREATE_REQUEST = 'CREATE_REQUEST';
-const DESTROY_REQUEST = 'DESTROY_REQUEST';
-const UPDATE_REQUEST = 'UPDATE_REQUEST';
-
-const LOAD_GAMES = 'LOAD_GAMES';
-const CREATE_GAME = 'CREATE_GAME';
-const DESTROY_GAME = 'DESTROY_GAME';
-const UPDATE_GAME = 'UPDATE_GAME';
-
 //*************************************************
 const usersReducer = (state = [], action) =>{
     if (action.type === LOAD_USERS){
@@ -34,23 +24,6 @@ const usersReducer = (state = [], action) =>{
     return state;
 }
 
-//*************************************************
-const requestsReducer = (state = [], action) =>{
-    if (action.type === LOAD_REQUESTS){
-        state = action.requests
-    }
-
-    return state;
-}
-
-//*************************************************
-const gamesReducer = (state = [], action) =>{
-    if (action.type === LOAD_GAMES){
-        state = action.games
-    }
-    return state;
-}
-
 
 //THUNKS****************************************
 
@@ -61,7 +34,7 @@ const _loadUsers = (users) =>{
     };
 };
 
-const loadUsers = () =>{
+export const loadUsers = () =>{
     return async(dispatch)=>{
         const users = (await axios.get('/api/users')).data;
         // console.log(users);
@@ -77,7 +50,7 @@ const _createUser = (user) =>{
 };
 
 // just the generic structure not actually working
-const createUser = (data, history)=>{
+export const createUser = (data, history)=>{
     return async(dispatch)=>{
         let user = (await axios.post('/api/users', { data })).data;
         dispatch(_createUser(user));
@@ -90,7 +63,7 @@ const createUser = (data, history)=>{
 // the records of the games they have played, not really sure
 const _destroyUser = user =>({ type: DESTROY_USER, user});
 
-const destroyUser = (user, history)=>{
+export const destroyUser = (user, history)=>{
     return async(dispatch)=>{
         await axios.delete(`/api/users/${user.id}`)
         dispatch(_destroyUser(user))
@@ -100,7 +73,7 @@ const destroyUser = (user, history)=>{
 
 const _updateUser = user =>({ type: UPDATE_USER, user});
 
-const updateUser = ( data )=>{
+export const updateUser = ( data )=>{
     return async(dispatch)=>{
         const user = (await axios.put(`/api/user/${id}`, { user })).data;
         dispatch(_updateUser(user));
@@ -108,39 +81,4 @@ const updateUser = ( data )=>{
 }
 
 
-
-//THUNKS****************************************
-
-const _loadRequests = (requests) =>{
-    return {
-        type: LOAD_REQUESTS,
-        requests
-    };
-};
-
-const loadRequests = () =>{
-    return async(dispatch)=>{
-        const requests = (await axios.get('/api/requests')).data;
-        dispatch(_loadRequests(requests));
-    }
-};
-
-//THUNKS****************************************
-
-const _loadGames = (games) =>{
-    return {
-        type: LOAD_GAMES,
-        games
-    };
-};
-
-const loadGames = () =>{
-    return async(dispatch)=>{
-        const games = (await axios.get('/api/games')).data;
-        dispatch(_loadGames(games));
-    }
-};
-
-
-// export default store;
-export { usersReducer, requestsReducer, gamesReducer, loadUsers, loadRequests, loadGames };
+export { usersReducer };
